+++ front/src/pages/trade/chart/candle/CandleBackground.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
-import { IChartItem } from '@src/recoil/chart';
-import userAtom, { IUser } from '@src/recoil/user/atom';
+import { IChartItem } from '@recoil/chart';
+import userAtom, { IUser } from '@recoil/user';
 import { TTheme, getMaxValue, getMinValue, getTextColor, getLegendColor, getText } from '../common';
 
 interface IProps {
