+++ front/src/pages/trade/chart/candle/CandleGraph.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useRef } from 'react';
 import { useRecoilValue } from 'recoil';
-import userAtom, { IUser } from '@src/recoil/user/atom';
-import { IChartItem } from '@src/recoil/chart/atom';
+import userAtom, { IUser } from '@recoil/user';
+import { IChartItem } from '@recoil/chart';
 
-import { IProps, TTheme, getMaxValue, getMinValue, CANDLE_GAP, getPriceColor } from '../common';
+import { IGraphComponentProps, TTheme, CANDLE_GAP, getMaxValue, getMinValue, getPriceColor } from '../common';
 import CandleBackground from './CandleBackground';
 import CandleLegend from './CandleLegend';
 
@@ -85,7 +85,7 @@ const drawCandles = ({
 	});
 };
 
-const CandleGraph = ({ chartData, crossLine, getYPosition }: IProps) => {
+const CandleGraph = ({ chartData, crossLine, getYPosition }: IGraphComponentProps) => {
 	const { theme } = useRecoilValue<IUser>(userAtom);
 	const candleGraphChartRef = useRef<HTMLCanvasElement>(null);
